Add a text renderer based on harfbuzz, rather than Pango+Cairo, which not only should be more widely available across platforms, but more improtantly, allows easier access to the raw integer coordinates of the font outlines, which take much less space in SVG paths than the floating-point coordinates I get from Pango+Cairo.

+++ build_view/make_artwork.sh
@@ -7,10 +7,15 @@
 DIR=$(dirname $(readlink -f $0))
 PRG="${DIR}/build_view.py"
 
+# Activate the virtual environment
+. ${DIR}/.venv/bin/activate
+
+COMPRESS=true
 OUTPUT=artwork
-while getopts o: name
+while getopts no: name
 do
     case $name in
+    n)  COMPRESS=false;;
     o)  OUTPUT="${OPTARG}";;
     ?)  printf "Usage: %s: [-o artwork-directory] <build-view.py flags>\n" $0
         exit 2;;
@@ -22,9 +27,11 @@ shift $(($OPTIND - 1))
 for K in vfx vfxsd sd1 sd132; do
   mkdir -p "${OUTPUT}/${K}"
   ${PRG} -l ${K} -io 'panel:' --text-paths --segment-paths --real-logos $@ > "${OUTPUT}/${K}/${K}.lay"
-  pushd "${OUTPUT}"
-  zip -r "${K}.zip" "${K}"
-  popd
-  rm "${OUTPUT}/${K}/${K}.lay"
-  rmdir "${OUTPUT}/${K}"
+  if [[ $COMPRESS = "true" ]]; then
+    pushd "${OUTPUT}"
+    zip -r "${K}.zip" "${K}"
+    popd
+    rm "${OUTPUT}/${K}/${K}.lay"
+    rmdir "${OUTPUT}/${K}"
+  fi
 done
